CRM-18522 tests to attempt to reproduce the reported sql error.
I could not reproduce the error through the UI alone. I had to basically hack the rules out in the db. I guess the config
arguably makes sense when configured in conjunction with the extension but that is not 100% clear.

However the code is a little less hacky & a lot better tested as a result of committing this